v6_wrapper/requirements.txt
openshift-client==1.0.13
requests
pyjwt
psycopg2==2.9.3
protobuf==3.20.*
tensorflow==2.8.0
nibabel==3.2.2
numpy==1.21.5
scipy==1.7.3
xnat==0.4.1